+++ dsl/src/main/kotlin/dev/vulnlog/dsl/VlBranchData.kt
@@ -10,3 +10,8 @@ public interface ReleaseVersionData {
     public val version: String
     public val releaseDate: LocalDate?
 }
+
+public interface InvolvedReleaseVersion {
+    public val affected: ReleaseVersionData?
+    public val upcoming: ReleaseVersionData?
+}