+++ dsl/src/main/kotlin/dev/vulnlog/dsl/VlReleasesDslRoot.kt
@@ -3,9 +3,6 @@ package dev.vulnlog.dsl
 import dev.vulnlog.dsl.definition.VlDslMarker
 
 public interface VlReleasesDslRoot {
-    public val branchToReleases: Map<ReleaseBranchData, List<ReleaseVersionData>>
-    public val reporters: List<ReporterData>
-
     /**
      * Define releases.
      *