+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add affected and fixed version information in report.
+
 ## [0.6.0] - 2025-03-15
 
 ### Added